requirements.txt
marshmallow==2.13.5
flask==0.12.2
flask-restplus
Flask-JWT-Extended==3.1.1
locustio==0.8a1